rename the packages
