+++ app/.env
@@ -1,0 +1,5 @@
+COHERE_API_KEY=cPfcdM4bmHI0c0ssSMcNOFXjgL04POWREj6o3z2v
+# REDIS_URL= redis://root:7RDCb8+GbgDABp0XPZyYVrTmQdkDq5If0Rke@theforevercanvas.com:6379
+REDIS_URL= redis://localhost:6379/0
+GOOGLE_API_KEY=AIzaSyCN3PenoV0dHzU3u2CHGFNizy5WP_C6DDQ
+LANGSMITH_API_KEY=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
